Extract popup error component
